fix(pool): stop a dust donate pinning liquidityIndex at its ceiling
Two dust donates on a never-minted leg pinned the index at type(uint64).max
for ~2 wei. raiseIndex then clamped forever while accrueLpFee kept booking
fees into liabilities, so 100% of that leg's fee stream became permanently
unclaimable. Value was destroyed, not captured.

Seed an unburnable floor at every liability-credit site (deposit,
swapLiability, donate, hookCreditYield) by splitting the caller's mint, so
liabilities can never return to dust. Widen liquidityIndex to uint96 with
INIT 1e18 in slot 1's spare bits, which makes the seed cheap enough that
the burn is negligible on expensive legs and the pin still costs millions
on cheap ones. Per-asset deadSeedPow10 normalises both tails.

Seeding skips rather than reverts on the degenerate books, matching
accrueLpFee: a revert inside settlement is a leg-wide DoS bought for gas.

adminRebaseIndexWidth migrates live legs: the index low bit slides 128->96
so old words read as oldIndex<<32. Legs storing 0 are legacy lazy-init, not
wiped, and are restored to the 1e12 base; genuinely wiped legs stay
terminal. Shift only, no rescale to the new base: rescaling an index
without its shares would multiply every claim by 1e6.

+++ evm/src/libraries/PoolAdmin.sol
@@ -182,7 +182,7 @@ library PoolAdmin {
     asset.vega = vega == 0 ? uint16(SC.BPS) : vega;
     asset.haircutSuppressor = uint16(SC.BPS);
     asset.pegB64 = M.encodeB64(SC.WAD, 18); // INTERNAL-mode default peg (1.0 base-per-asset)
-    asset.liquidityIndex = uint64(C.LIQUIDITY_INDEX_INIT); // explicit: 0 now means "wiped", not "unset"
+    asset.liquidityIndex = uint96(C.LIQUIDITY_INDEX_INIT); // explicit: 0 now means "wiped", not "unset"
     asset.lastUpdate = uint32(block.timestamp); // A2-1: seed so first decay enable has no retroactive dt
 
     asset.anchor = t == $.baseToken ? address(0) : $.baseToken;
